Removed external status LED from the schematic.
Rex pointed out this creates more of a current load on the pin,
raising it from about 15 mA to 30 mA, above the recommended limit
for a single pin on the Pro Mini.  Since it adds little value
beyond the on-board LED, and since we're going to need more board
space for other components, I just deleted it.