Improved the laser switching.
The pull-down resistor was unnecessary and complicated computing the
base resistor by creating a voltage divider.

I increased the value of the base resistor to 1.5 kOhms to further
limit the current into the base.  The target is now 2.5 mA for a
0.89 V base-emitter drop, based on my understanding of the 2N3904
datasheet and the 25 mA required by the laser diode.