Changed interface between MCU and the audio player.
Decided to use SoftwareSerial instead of the UART so that the
console remains available for debugging.  Since SoftwareSerial
blocks interrupts, writes to the audio player will be limited to
starting sound effects on a trigger before starting the actual
effect.  Tesponses from the audio player will be ignored, and the
BUSY signal from the player can be monitored to detect when a
track starts or stops.

Upped the values in the voltage divider to reduce the current to
1.5 mA.