Updated bill of materials to switch Arduino Pro Mini version.
The cheaper ones from Hiletgo were unreliable to the voltage regulators.
Ended up switching to the actual SparkFun ones.  Unfortunately, those
may still take a few days to arrive, and the up the kit price by $4.

Fortunately, I figured out how to continue to develop the software
using an old Pro Mini with a blown voltage regulator:  I clipped off the
RAW pin on the Pro Mini and keep it connected to the FTDI cable.  That
way the Pro Mini (and other things powered by its VCC) get their 5V from
the USB hub, bypassing the voltage regulator.  Since we need the 12V for
the fan, that can be powered from the jack on the board (as intended).