Prevented serial audio message receiving from getting stuck in a loop.
When the handling of a message caused another message to arrive
immediately, it could get stuck in the `update()` indefinitely.  Now
it'll process only the pending bytes available when `update()` is
first called.  Bytes that arrive during processing will remain in the
stream's queue until the next call to update.

For this project, this was important because getting stuck in the
audio module's update made emergency stop and suppress inoperable.