Added holes to deflectors to make removing the mirror easier.
Generally, the mirrors should be glued in.  But sometimes you want
to snap it in first to test the fit.  If it's snug, it can be
difficult to remove the mirror.  Now you can insert a small tool,
like the top of a screwdriver, from the back to pop the mirror
back out.