Made the laser mount mast more substantial.
Several laser mounts snapped off at the base of the mast during
shipping.  The original design purposely made the mast narrow to
minimize the size of the shadow it might cast.  But it's unusual
to select angles that put the mast in the path of the beam, so
the thin design was unnecessary.  I've updated it to a tapered
cylinder.  This should also reduce printing problems due to a
lack of support for the mount.